+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 9.7.0
+
+* Added `FlAutomaticKeepAliveWrapper` and `FlAutomaticKeepAliveWrapperState`
+* Modify `DecoratorEntry` to `DecoratedPendant`
+* Remove `DecoratorPositioned.outermost`
+* Modify `DecoratorPositioned` to `DecoratedPendantPosition`
+* `DecoratedPendant` adds the `maintainSize` property to determine whether to maintain the size
+* Modify `DecoratorBoxState` to `FlDecoratedBoxState`
+* Modify `DecoratorBox` to `FlDecoratedBox`
+* Added `FlBoxDecoration` to `FlDecoratedBox` and `FlDecoratedBoxState`
+* Modify `.toDecoratorEntry()` to `.toDecoratedPendant()`
+
 ## 9.6.0
 
 * Remove the `DropdownMenusButton` component, please use the `MultiPopupMenuButton` component